feat: add @next/mdx and migrate to mjs file ext
