+++ apps/web/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@formatjs/intl-localematcher": "^0.5.4",
+    "@next/mdx": "^14.2.4",
     "@notionhq/client": "^2.2.15",
     "@tanstack/react-query": "^5.45.1",
     "@vercel/analytics": "^1.2.2",
